now tests for points not moving in frame 0, and they do.
