+++ templates/analyze.html
@@ -49,9 +49,10 @@ line-height:0px;
 
 <div id='template' >
   <table class='float-container'>
-    <tr><td class='message'>Loading frame...</td><td>&nbsp;</td></tr>
+    <tr><td class='status_message'>Loading frame...</td><td>&nbsp;</td></tr>
     <tr> <!-- First row of analysis table: the movie being tracked-->
-      <td> <canvas id="canvas-id" width='320' height='240'> </canvas><br>
+      <td>
+        <canvas id="canvas-id" width='320' height='240'> </canvas></br>
         <form action="javascript:void(0)" >
           <label for="zoom">zoom:</label>
           <select id="zoom-id" name="zoom" class='zoom'>
