+++ tests/Makefile
@@ -4,7 +4,7 @@ extract_and_upload:
 
 # https://stackoverflow.com/questions/14405063/how-to-see-normal-print-output-created-during-pytest-run
 bt:
-	cd ..;python3 -m pytest tests/tracker_test.py -k test_blocktrack -v --log-cli-level=DEBUG
+	cd ..;python3 -m pytest tests/tracker_test.py -k test_movie_tracking -v --log-cli-level=DEBUG
 
 pytest:
 	@echo use pytest-debug if make is run from here
